requirements.txt
aiohttp==3.6.3
argh==0.26.2
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
cffi==1.14.3
chardet==3.0.4
discord.py==1.5.1
idna==2.10
idna-ssl==1.1.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pathtools==0.1.2
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
python-dotenv==0.15.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
watchdog==0.10.3
wrapt==1.12.1
yarl==1.5.1
